méthodes gestion de produits et catégories pour l'admin

+++ test/src/main/java/fr/adaming/dao/IProduitDao.java
@@ -8,7 +8,7 @@ import fr.adaming.model.Produit;
 public interface IProduitDao {
 
 	public List<Produit> getAllProduit();
-	public Produit getProduitById(int id);
+	public Produit getProduitById(long id);
 	public Produit addProduit(Produit produit);
 	public void updateProduit(Produit produit);
 	public void deleteProduit(Produit produit);
