+++ test/src/main/java/fr/adaming/service/ProduitServiceImpl.java
@@ -32,7 +32,7 @@ public class ProduitServiceImpl implements IProduitService {
 	}
 
 	@Override
-	public Produit getProduitById(int id) {
+	public Produit getProduitById(long id) {
 		return produitDao.getProduitById(id);
 	}
